Allow emscripten-releases-tags/aliases to work be used without SDK deps
This change means that aliases such as `latest` and `3.10.2` can be
used to install jsut the emscipten-releases packages and not the full
SDK with all the dependencies.

This allows uses who do not want to install dependencies such as node,
python or java to still use this package.

These packages were already available to install on their own via
their full names such as:

```
$ ./emsdk install releases-upstream-48ce0b44015d0182fc8c27aa9fbc0a4474b55982-64bi
```

Now these base packages can also be install via:

```
$ ./emsdk install latest-base
```

Fixes: #1142